refactor(trainer): moved all training related files into optimum/neuron/trainers/
