+++ optimum/neuron/trainers/__init__.py
@@ -13,8 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .transformers import NeuronTrainer, Seq2SeqNeuronTrainer
-from .sft_trainer import NeuronSFTTrainer
 from .orpo_trainer import NeuronORPOTrainer
+from .sft_trainer import NeuronSFTTrainer
 from .training_args import NeuronTrainingArguments, Seq2SeqNeuronTrainingArguments
-from .trl_utils import NeuronSFTConfig, NeuronORPOConfig
+from .transformers import NeuronTrainer, Seq2SeqNeuronTrainer
+from .trl_utils import NeuronORPOConfig, NeuronSFTConfig
