+++ optimum/neuron/trainers/orpo_trainer.py
@@ -13,119 +13,39 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import contextlib
-import dataclasses
-import functools
 import inspect
-import math
-import os
-import shutil
-import sys
-import time
 import warnings
 from collections import defaultdict
-from functools import wraps
-from typing import Any, Callable
+from typing import Callable
 
-import datasets
-import numpy as np
 import torch
 import torch_xla.core.xla_model as xm
-import torch_xla.debug.metrics as met
-import torch_xla.runtime as xr
-from accelerate import __version__ as accelerate_version
-from accelerate.utils import AutocastKwargs, DataLoaderConfiguration
-from neuronx_distributed.pipeline import NxDPPModel
-from packaging import version
 from torch import nn
 from torch.utils.data import Dataset
 from transformers import (
     AutoModelForCausalLM,
-    AutoTokenizer,
     DataCollator,
-    DataCollatorForLanguageModeling,
     PreTrainedModel,
     PreTrainedTokenizerBase,
-    Seq2SeqTrainer,
     Trainer,
-    TrainingArguments,
     is_wandb_available,
 )
-from transformers.debug_utils import DebugOption, DebugUnderflowOverflow
-from transformers.integrations import hp_params
-from transformers.modeling_utils import unwrap_model
-from transformers.trainer import (
-    OPTIMIZER_NAME,
-    SCHEDULER_NAME,
-    TRAINER_STATE_NAME,
-    TRAINING_ARGS_NAME,
-)
-from transformers.trainer_callback import TrainerCallback, TrainerState
-from transformers.trainer_pt_utils import (
-    IterableDatasetShard,
-    find_batch_size,
-    get_dataloader_sampler,
-    nested_concat,
-    nested_numpify,
-    reissue_pt_warnings,
-)
+from transformers.trainer_callback import TrainerCallback
 from transformers.trainer_utils import (
-    PREFIX_CHECKPOINT_DIR,
     EvalLoopOutput,
-    EvalPrediction,
-    HPSearchBackend,
-    PredictionOutput,
-    SaveStrategy,
-    TrainOutput,
-    denumpify_detensorize,
-    has_length,
-    speed_metrics,
-)
-from transformers.utils import (
-    WEIGHTS_NAME,
-    is_accelerate_available,
-    is_sagemaker_mp_enabled,
 )
 
-from optimum.utils import logging
-
-from .accelerate import NeuronAccelerator, NeuronDistributedType, NeuronPartialState
-from .cache.hub_cache import hub_neuronx_cache, synchronize_hub_cache
-from .cache.training import patch_neuron_cc_wrapper
-from .peft import NeuronPeftModel, get_peft_model
-from .training_args import NeuronTrainingArguments
-from .utils import (
-    is_trl_available,
-    patch_within_function,
-)
-from .utils.cache_utils import (
-    get_hf_hub_cache_repos,
-    get_neuron_cache_path,
-    has_write_access_to_repo,
-)
-from .utils.import_utils import is_peft_available
-from .utils.misc import is_main_worker, is_precompilation
-from .utils.require_utils import requires_torch_neuronx
-from .utils.training_utils import (
-    get_model_param_count,
-    is_main_worker_for_metrics,
-    is_main_worker_for_metrics_method,
-    patch_generation_mixin_to_neuron_generation_mixin,
-    skip_first_batches,
-)
-from .utils.trl_utils import NeuronSFTConfig
-
-from .transformers import Trainer
-from .base import _TrainerForNeuron
+from ..accelerate import NeuronPartialState
+from ..peft import NeuronPeftModel, get_peft_model
 from ..utils import (
     is_trl_available,
-    patch_within_function,
 )
+from ..utils.import_utils import is_peft_available
+from .base import TRL_VERSION, _TrainerForNeuron
 
-from .trl_utils import NeuronSFTConfig
 
 if is_trl_available():
-    from trl import ORPOConfig, ORPOTrainer, SFTConfig, SFTTrainer
+    from trl import ORPOConfig, ORPOTrainer
 else:
     class ORPOConfig:
         pass
